% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{extract_text}
\alias{extract_text}
\title{Extract text from photo with GoogleVision API}
\usage{
extract_text(img_path)
}
\arguments{
\item{img_path}{a file path or URL to an image file of type jpg, tiff, or png}
}
\description{
\code{extract_text} Calls GoogleVision API with TEXT_DETECTION feature
}
\details{
This is a function that simply calls the GoogleVision API and passes a photo URL
The only argument is \code{imgPath} which should be a file path to an image file.
}
\examples{
\dontrun{
img_file_location <- "./vignettes/data/raw_data/example_images/IMG_0797.JPG"
image_text <- extract_text(img_file_location)
}
}
