% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_box_coords}
\alias{get_box_coords}
\title{Create bounding box coordinates}
\usage{
get_box_coords(img, image_text)
}
\description{
\code{get_box_coords} Strips bounding box coordinates from API response
and returns a list includng the image bounding box as \code{bbox} and
extracted text bounding boxes as \code{bbox_coords}
}
