% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{ggplot_bbox_coords}
\alias{ggplot_bbox_coords}
\title{Create ggplot object of bounding boxes}
\usage{
ggplot_bbox_coords(img, bbox_coords, bbox)
}
\description{
\code{ggplot_bbox_coords} Builds ggplot object from image, tags, and bounding boxes
}
