% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{make_tags}
\alias{make_tags}
\title{Convert API response to tags}
\usage{
make_tags(image_text)
}
\arguments{
\item{image_text}{a data.frame of extracted text as returned from \code{extract_text}}
}
\description{
\code{make_tags} Takes the response from GoogleVision API and returns tags for metadata
}
