% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{plot_results}
\alias{plot_results}
\title{Plots text and bounding boxes on photo}
\usage{
plot_results(img_path, image_text)
}
\arguments{
\item{img_path}{a file path or URL to an image file of type jpg, tiff, or png}

\item{image_text}{a data.frame of extracted text as returned from \code{extract_text} using the same image as \code{img_path}}
}
\description{
\code{plot_results} Utilizes ggplot2 to plot extracted text and bounding boxes on image
}
\examples{
\dontrun{
img_file_location <- "./vignettes/data/raw_data/example_images/IMG_0797.JPG"
image_text <- extract_text(img_file_location)
img_plot <- plot_results(img_file_location, image_text)
img_plot
}
}
