% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{write_XMP}
\alias{write_XMP}
\title{Write metadata tags to image}
\usage{
write_XMP(img_path, img_tags)
}
\arguments{
\item{img_path}{a file path or URL to an image file of type jpg, tiff, or png}

\item{img_tags}{a character vector of metadata tags}
}
\description{
\code{extract_text} Writes XMP metadata tags to image using exiftools
}
\seealso{
\code{\link{extract_text}} and \code{\link{make_tags}}
}
